updater: Refactor RepoManager in preparation of handling user confirmation
The RepoManager is now called UpdateManager as it really manages updates rather than repos.

The UserActionRequiredListener has been moved into its own file to keep PackageInstaller small and focused.

The installation part of UpdateManager was factored out into a InstallManager which will be managing user install confirmations as well.

The UpdaterManager is now able to return a UpdateAppsResult which is more flexible than a single boolean.

Change-Id: I3e37502f337597e5f601ce0a5c0acb8ee5837867