installer: Don't crash when user presses next and has no apps selected
We would start the foreground service even with nothing to install and stop it right away before we could call the startForeground message, then the system would crash us, because we haven't called that method. Not starting the service at all seems to be a good solution to this.

Issue: calyxos#1603
Change-Id: I40ef715be14f533ead627a959951715bb8745ae0